// src/interfaces/article.ts
// Generated by https://quicktype.io

export interface Article {
  totalArticles: number;
  articles: ArticleElement[];
}

export interface ArticleElement {
  title: string;
  description: string;
  content: string;
  url: string;
  image: string;
  publishedAt: Date;
  source: Source;
}

export interface Source {
  name: string;
  url: string;
}